requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
mysqlclient==2.0.3
Pillow==8.3.2
PyJWT==1.7.1
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
